--- pages/landingPage/AboutUsPage.js
@@ -1,14 +1,0 @@
-import { StyledMain } from "../../components/styles";
-
-const AboutUsPage = () => {
-  return (
-    <StyledMain>
-      <h2>About us</h2>
-      <article>
-        <p>this is the about us page</p>
-      </article>
-    </StyledMain>
-  );
-};
-
-export default AboutUsPage;
