--- pages/landingPage/PantePersonSelectorFiles/PantePersonData.js
@@ -1,5 +1,0 @@
-//BERGEN
-
-const people = ["Knut", "Morthen", "Øystein", "Øystein 2", "Emil sin tur", "Gønnaren"];
-
-export default people;
