--- pages/landingPage/PantePersonSelectorFiles/PantePersonSelectorStyles.js
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const PanteSelectorTitle = styled.h2`
-  font-size: 2rem;
-`;
-
-export const RandomPerson = styled.p`
-  font-size: 3rem;
-`;
