--- pages/landingPage/StudentProjectsPage.js
@@ -1,14 +1,0 @@
-import { StyledMain } from "../../components/styles";
-
-const StudentProjectsPage = () => {
-  return (
-    <StyledMain>
-      <h2>Student Projects</h2>
-      <article>
-        <p>Student projects are to be filled inn here</p>
-      </article>
-    </StyledMain>
-  );
-};
-
-export default StudentProjectsPage;
