--- pages/landingPage/aboutUsPage/AboutUsPage.js
@@ -1,1 +1,0 @@
-
